Correctly set node attributes
If the current node is part of an array, like when there is multiple
children with the same name, only the first one had its attributes set,
due to https://github.com/hapijs/faketoe/blob/master/lib/index.js#L96.

Now, if it's in an array, the last item of the array is given the
attributes.

I've updated the test and the fixture to reflect that.

Close #24